Updated instructions to enable auto subdomains
The current instructions say to click the "Edit domain" button but there is no such button in the UI. The correct button to click is "Manage subdomains" and then you can enable the feature from there. Also I added a warning that says this is only possible for root domains.